fix: always redo journal files and append targets
For some reason after switching to Python for this file, redo-ifchange
causes a message "not marked as generated -- not redoing" to appear
for all the journal files generated. I don't know how to fix that,
and at any rate, it's probably better to always redo them each time.
If there's nothing new, they don't get written anyway, and with one file
per year per account, it's not going to be a major slowdown.